Sync with QtCreator
Require a QMimedatabase parameter in the constructor of QMimeDatabaseBuilder.
Removed the direct usage of QMimeDatabaseBuilder in BaseMimeTypeParser.
Made XML tags in MIME files accessible to the outside.
Readded a bunch of code for QtCreator which will be kept internal though from now on.
Finished David's work with QMimeTypeData::matchesFileBySuffix().
Added a new constructor to QMimeDatabase and QMimeDatabasebuilder with a QMimeDatabasePrivate argument to support private databases outside staticMimeDataBase() in QtCreator. The constructor will thus delete the d pointer if it is not set to staticMimeDataBase().
Export QMimeDatabaseBuilder and QMimeTypeData.
Added a constructor to QMimeTypeData from QMimeType
Declare default constructor and constructor from QMimeTypeData of QMimeType as public.
Updated the documentation.