Add findMimeTypesByFileName which returns a list of mimetypes.
This is needed in order to implement delayed mimetype determination.
If, based on the filename only, we get 0 or multiple mimetypes, then
we know findByData will be necessary later. If we get exactly 1
mimetype, then that is authoritative and we won't need findByData.

Change-Id: Ib22edb512fb45099268aafd92500109849c331c4
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Wolf-Michael Bolle <wolf-michael.bolle@nokia.com>