Adapt to Qt4 QSKIP
From the C++ API changes at
http://qt-project.org/doc/qt-5.0/qtdoc/sourcebreaks.html#changes-to-qt-test

"The "mode" parameter in QSKIP macro is removed. This parameter caused
problems in test metrics calculation because of the SkipAll mode, which
hid information about the skipped test data. Calling QSKIP in a test
function now behaves like SkipSingle, which is skipping a non-data-driven
test function or skipping only the current data row of a data-driven test
function. Every skipped data row is now reported in the test log."

Change-Id: I6d07fb13f5949dda52a89438d84c86dcc54f8c6c
Signed-off-by: Luís Pereira <luis.artur.pereira@gmail.com>
Reviewed-by: David Faure (KDE) <faure@kde.org>